+++ tests/test_logic.py
@@ -291,3 +291,38 @@ def test_error_messages(cdataset):
         rule = rule_class(bmds.constants.BIN_FAILURE, threshold=1)
         _, msg = rule.apply_rule(cdataset, outputs)
         assert msg == expected
+
+
+def test_parsimonious_recommendation(reduced_cdataset):
+    session = bmds.BMDS.latest_version(bmds.constants.CONTINUOUS, dataset=reduced_cdataset)
+    session.add_default_models()
+    session.execute()
+    session.recommend()
+
+    # confirm recommended model is Hill on basis of lowest BMDL
+    assert session.recommended_model.model_name == 'Hill'
+    assert session.recommended_model.recommended_variable == 'BMDL'
+
+    # exponential M2 < exponential M3
+    models = [
+        model for model in session.models
+        if model.output['AIC'] == 158.9155
+    ]
+    assert len(models) == 2
+    assert bmds.Recommender._get_parsimonious_model(models).name == 'Exponential-M2'
+
+    # exponential M4 < exponential M5
+    models = [
+        model for model in session.models
+        if model.output['AIC'] == 155.5369
+    ]
+    assert len(models) == 2
+    assert bmds.Recommender._get_parsimonious_model(models).name == 'Exponential-M4'
+
+    # linear < (polynomial, power)
+    models = [
+        model for model in session.models
+        if model.output['AIC'] == 159.370875
+    ]
+    assert len(models) == 6
+    assert bmds.Recommender._get_parsimonious_model(models).name == 'Linear'